UI/requirements.txt
accelerate==0.26.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
asttokens==2.0.5
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
bitarray==2.9.2
blis==0.7.11
bs4==0.0.2
build==1.1.1
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cloudpathlib==0.16.0
cohere==4.45
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.4
cryptography==42.0.1
cymem==2.0.8
dataclasses-json==0.6.3
datasets==2.16.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.8.0rc2
Deprecated==1.2.14
dill==0.3.7
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dj-rest-auth==5.0.2
Django==5.0.1
django-cors-headers==4.3.1
django-db-file-storage==0.5.6.1
django-stubs==4.2.6
django-stubs-ext==4.2.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
entrypoints==0.4
evaluate==0.4.1
exceptiongroup==1.2.0
executing==0.8.3
fastapi==0.110.0
fastavro==1.9.3
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2023.10.0
google-auth==2.28.1
googleapis-common-protos==1.62.0
gptcache==0.1.43
greenlet==3.0.3
grpcio==1.62.0
guidance==0.0.64
gunicorn==20.0.4
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.2
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.1.2
interegular==0.3.3
ipykernel==6.9.2
ipython==8.1.1
jedi==0.18.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-ng==1.6.1
jsonpointer==2.4
kubernetes==29.0.0
langchain==0.1.0
langchain-community==0.0.16
langchain-core==0.1.17
langchain-text-splitters==0.0.1
langcodes==3.3.0
langsmith==0.0.84
llama-index==0.9.35
lm-format-enforcer==0.4.3
load-dotenv==0.1.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib-inline==0.1.3
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msal==1.26.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
mypy==1.5.1
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
oauthlib==3.2.2
onnx==1.15.0
onnxruntime==1.16.3
openai==1.7.1
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
optimum==1.16.1
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.1.4
parso==0.8.3
pathlib2==2.3.7.post1
pdfreader==0.1.12
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.1.0
ply==3.11
posthog==3.5.0
preshed==3.0.9
prompt-toolkit==3.0.28
protobuf==4.25.2
psutil==5.9.8
psycopg2-binary==2.9.5
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.11.2
pygtrie==2.5.0
PyJWT==2.8.0
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pyparsing==3.1.1
pypdf==4.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3.post1
pywin32==303
PyYAML==6.0.1
rank-bm25==0.2.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.25
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.2.0
starlette==0.36.3
sympy==1.12
tenacity==8.2.3
thinc==8.2.2
threadpoolctl==3.2.0
tiktoken==0.5.2
timm==0.9.12
tokenizers==0.15.0
tomli==2.0.1
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.36.2
typer==0.9.0
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.27.1
vue==0.0.1
waitress==3.0.0
wasabi==1.1.2
watchfiles==0.21.0
wcwidth==0.2.5
weasel==0.3.4
websocket-client==1.7.0
websockets==12.0
whitenoise==6.6.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
